+++ src/test/java/kitchenpos/menu/service/MenuGroupServiceTest.java
@@ -1,8 +1,9 @@
-package kitchenpos.menu;
+package kitchenpos.menu.service;
 
 import kitchenpos.application.MenuGroupService;
 import kitchenpos.dao.MenuGroupDao;
 import kitchenpos.domain.MenuGroup;
+import kitchenpos.menu.MenuFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
