+++ src/test/java/kitchenpos/order/service/TableServiceTest.java
@@ -1,10 +1,11 @@
-package kitchenpos.order;
+package kitchenpos.order.service;
 
 import kitchenpos.application.TableService;
 import kitchenpos.dao.OrderDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
+import kitchenpos.order.OrderFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
